requirements.txt
GitPython 
pytest
# ensure `dotnet` (SDK 8) & `ollama` are on your PATH
